+++ src/stories/forms/DrugSearch.stories.jsx
@@ -1,12 +1,11 @@
-
 import React from 'react';
 import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
 
-import { initialState as medInitialState } from '../utils/store/medSlice';
-import medSliceReducer from '../utils/store/medSlice';
+import { initialState as medInitialState } from '../../utils/store/medSlice';
+import medSliceReducer from '../../utils/store/medSlice';
 
-import DrugSearch from '../components/DrugSearch';
+import DrugSearch from '../../components/DrugSearch';
 
 const store  = configureStore({
   medInitialState,
